api/requirements.txt
Flask==3.1.0
Flask_Cors==5.0.0
pandas==2.2.3
protobuf==5.29.3
python-dotenv==1.0.1
gunicorn>=20.1.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3